fix: split constant usage to minimize file size of inpage
