+++ src/extension/background-script/actions/nostr/decryptOrPrompt.ts
@@ -1,3 +1,4 @@
+import { USER_REJECTED_ERROR } from "~/common/constants";
 import utils from "~/common/lib/utils";
 import state from "~/extension/background-script/state";
 import i18n from "~/i18n/i18nConfig";
@@ -53,7 +54,7 @@ const decryptOrPrompt = async (message: MessageDecryptGet) => {
 
         return { data: response };
       } else {
-        return { error: "User rejected" };
+        return { error: USER_REJECTED_ERROR };
       }
     }
   } catch (e) {
