+++ src/extension/background-script/actions/nostr/encryptOrPrompt.ts
@@ -1,3 +1,4 @@
+import { USER_REJECTED_ERROR } from "~/common/constants";
 import utils from "~/common/lib/utils";
 import state from "~/extension/background-script/state";
 import i18n from "~/i18n/i18nConfig";
@@ -51,7 +52,7 @@ const encryptOrPrompt = async (message: MessageEncryptGet) => {
 
         return { data: response };
       } else {
-        return { error: "User rejected" };
+        return { error: USER_REJECTED_ERROR };
       }
     }
   } catch (e) {
