+++ src/extension/background-script/state.ts
@@ -2,8 +2,8 @@ import merge from "lodash.merge";
 import pick from "lodash.pick";
 import browser from "webextension-polyfill";
 import createState from "zustand";
-import { DEFAULT_SETTINGS } from "~/common/constants";
 import { decryptData } from "~/common/lib/crypto";
+import { DEFAULT_SETTINGS } from "~/common/settings";
 import { isManifestV3 } from "~/common/utils/mv3";
 import { Migration } from "~/extension/background-script/migrations";
 import type { Account, Accounts, SettingsStorage } from "~/types";
