+++ tests/fixtures/settings.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_SETTINGS } from "~/common/constants";
+import { DEFAULT_SETTINGS } from "~/common/settings";
 import type { SettingsStorage } from "~/types";
 
 export const settingsFixture: SettingsStorage = { ...DEFAULT_SETTINGS };
